+++ share_your_route_front/lib/classes/destinations.dart
@@ -1,12 +1,14 @@
 class Destinations {
   final String name;
   final double duration;
+  final String image;
   final String description;
   final String startTime;
 
   Destinations(
       {required this.name,
       required this.duration,
+      required this.image,
       required this.description,
       required this.startTime});
 }
