+++ share_your_route_front/lib/login_page.dart
@@ -1,0 +1,116 @@
+import 'package:flutter/material.dart';
+import 'package:share_your_route_front/homePage.dart';
+
+class LoginPage extends StatelessWidget {
+  const LoginPage({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return MaterialApp(
+      theme: ThemeData(useMaterial3: true),
+      home: const Login(),
+    );
+  }
+}
+
+class Login extends StatefulWidget {
+  const Login({super.key});
+
+  @override
+  State<Login> createState() => LoginState();
+}
+
+class LoginState extends State<Login> {
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      backgroundColor: Colors.white,
+      body: Center(
+        child: SingleChildScrollView(
+          child: Padding(
+            padding: const EdgeInsets.all(16.0),
+            child: Column(
+              mainAxisAlignment: MainAxisAlignment.center,
+              mainAxisSize: MainAxisSize.min,
+              children: <Widget>[
+                Padding(
+                  padding: const EdgeInsets.only(top: 60.0),
+                  child: Center(
+                    // ignore: sized_box_for_whitespace
+                    child: Container(
+                      width: 200,
+                      height: 150,
+                      child: Image.asset('asset/images/logo.png'),
+                    ),
+                  ),
+                ),
+                const Padding(
+                  padding: EdgeInsets.symmetric(horizontal: 15),
+                  child: TextField(
+                    decoration: InputDecoration(
+                      border: OutlineInputBorder(),
+                      labelText: 'Email',
+                      hintText: 'Email válido como: abc@gmail.com',
+                    ),
+                  ),
+                ),
+                const Padding(
+                  padding: EdgeInsets.only(
+                    left: 15.0,
+                    right: 15.0,
+                    top: 15,
+                  ),
+                  child: TextField(
+                    obscureText: true,
+                    decoration: InputDecoration(
+                      border: OutlineInputBorder(),
+                      labelText: 'Contraseña',
+                      hintText: 'Ingrese una contraseña segura',
+                    ),
+                  ),
+                ),
+                TextButton(
+                  onPressed: () {
+                    //TODO FORGOT PASSWORD SCREEN GOES HERE
+                  },
+                  child: const Text(
+                    'Olvidaste tu contraseña?',
+                    style: TextStyle(
+                      color: Color.fromRGBO(37, 60, 89, 1),
+                      fontSize: 15,
+                    ),
+                  ),
+                ),
+                Container(
+                  height: 50,
+                  width: 250,
+                  decoration: BoxDecoration(
+                    color: const Color.fromRGBO(37, 60, 89, 1),
+                    borderRadius: BorderRadius.circular(20),
+                  ),
+                  child: TextButton(
+                    onPressed: () {
+                      Navigator.push(
+                        context,
+                        MaterialPageRoute(builder: (_) => const HomePage()),
+                      );
+                    },
+                    child: const Text(
+                      'Iniciar sesión',
+                      style: TextStyle(
+                        color: Colors.white,
+                        fontSize: 25,
+                      ),
+                    ),
+                  ),
+                ),
+                const SizedBox(height: 130),
+                const Text('Usuario nuevo? Crea una cuenta'),
+              ],
+            ),
+          ),
+        ),
+      ),
+    );
+  }
+}
