+++ share_your_route_front/lib/main.dart
@@ -1,108 +1,54 @@
 // ignore_for_file: prefer_const_constructors, prefer_const_literals_to_create_immutables
 
+import 'dart:async';
+
 import 'package:flutter/material.dart';
-import 'package:share_your_route_front/home.dart';
-import 'package:share_your_route_front/itinerarioRuta.dart';
-import 'package:share_your_route_front/createRoute2.dart';
+import 'package:share_your_route_front/login_page.dart';
 
 void main() => runApp(MaterialApp(
       debugShowCheckedModeBanner: false,
-      home: LoginPage(),
+      home: MainPage(),
     ));
 
-class LoginPage extends StatefulWidget {
+class MainPage extends StatefulWidget {
   @override
-  _LoginPageState createState() => _LoginPageState();
+  _MainPageState createState() => _MainPageState();
 }
 
-class _LoginPageState extends State<LoginPage> {
+class _MainPageState extends State<MainPage> {
+  @override
+  void initState() {
+    super.initState();
+    Timer(Duration(seconds: 3), () {
+      Navigator.pushReplacement(
+        context,
+        MaterialPageRoute(builder: (context) => LoginPage()),
+      );
+    });
+  }
+
   @override
   Widget build(BuildContext context) {
     return Scaffold(
       backgroundColor: Colors.white,
       body: Center(
-        child: SingleChildScrollView(
-          child: Padding(
-            padding: const EdgeInsets.all(16.0),
-            child: Column(
-              mainAxisAlignment: MainAxisAlignment.center,
-              mainAxisSize: MainAxisSize.min,
-              children: <Widget>[
-                Padding(
-                  padding: const EdgeInsets.only(top: 60.0),
-                  child: Center(
-                    child: Container(
-                      width: 200,
-                      height: 150,
-                      child: Image.asset('asset/images/logo.png'),
-                    ),
-                  ),
+        child: Column(
+          mainAxisAlignment: MainAxisAlignment.center,
+          children: [
+            Padding(
+              padding: const EdgeInsets.only(top: 60.0),
+              child: Center(
+                child: Container(
+                  width: 200,
+                  height: 150,
+                  child: Image.asset('asset/images/logo.png'),
                 ),
-                Padding(
-                  padding: EdgeInsets.symmetric(horizontal: 15),
-                  child: TextField(
-                    decoration: InputDecoration(
-                      border: OutlineInputBorder(),
-                      labelText: 'Email',
-                      hintText: 'Email válido como: abc@gmail.com',
-                    ),
-                  ),
-                ),
-                Padding(
-                  padding: const EdgeInsets.only(
-                    left: 15.0,
-                    right: 15.0,
-                    top: 15,
-                  ),
-                  child: TextField(
-                    obscureText: true,
-                    decoration: InputDecoration(
-                      border: OutlineInputBorder(),
-                      labelText: 'Contraseña',
-                      hintText: 'Ingrese una contraseña segura',
-                    ),
-                  ),
-                ),
-                TextButton(
-                  onPressed: () {
-                    //TODO FORGOT PASSWORD SCREEN GOES HERE
-                  },
-                  child: Text(
-                    'Olvidaste tu contraseña?',
-                    style: TextStyle(
-                      color: Color.fromRGBO(37, 60, 89, 1),
-                      fontSize: 15,
-                    ),
-                  ),
-                ),
-                Container(
-                  height: 50,
-                  width: 250,
-                  decoration: BoxDecoration(
-                    color: Color.fromRGBO(37, 60, 89, 1),
-                    borderRadius: BorderRadius.circular(20),
-                  ),
-                  child: TextButton(
-                    onPressed: () {
-                      Navigator.push(
-                        context,
-                        MaterialPageRoute(builder: (_) => HomePage()),
-                      );
-                    },
-                    child: Text(
-                      'Iniciar sesión',
-                      style: TextStyle(
-                        color: Colors.white,
-                        fontSize: 25,
-                      ),
-                    ),
-                  ),
-                ),
-                SizedBox(height: 130),
-                Text('Usuario nuevo? Crea una cuenta'),
-              ],
+              ),
             ),
-          ),
+            CircularProgressIndicator(color: Color.fromRGBO(37, 60, 89, 1)),
+            SizedBox(height: 20),
+            Text('Cargando...'),
+          ],
         ),
       ),
     );
